add chart points initializer

+++ src/main/java/model/Chart.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.util.Arrays;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -11,8 +13,8 @@ public class Chart extends BaseEntity {
 
     private Point[] points;
 
-    private int suggestedMinX;
-    private int suggestedMaxX;
-    private int suggestedMinY;
-    private int suggestedMaxY;
+    private double suggestedMinX;
+    private double suggestedMaxX;
+    private double suggestedMinY;
+    private double suggestedMaxY;
 }
